+++ Formatting/Hashtable.format.ps1
@@ -1,0 +1,3 @@
+Write-FormatView -TypeName Hashtable -Name Hashtable -Action {
+    $_ | Format-Hashtable -Depth $FormatEnumerationLimit
+}