+++ Transpilers/Inline/Inline.CSS.psx.ps1
@@ -40,6 +40,7 @@ MyClass {
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline)]
+[Management.Automation.CommandInfo]
 $CommandInfo,
 
 # A dictionary of parameters.