+++ Transpilers/Inline/Inline.Markdown.psx.ps1
@@ -43,6 +43,7 @@
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline)]
+[Management.Automation.CommandInfo]
 $CommandInfo,
 
 # A dictionary of parameters.