+++ Transpilers/Inline/Inline.OpenSCAD.psx.ps1
@@ -46,6 +46,7 @@ if (Shape == "circle") {
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline)]
+[Management.Automation.CommandInfo]
 $CommandInfo,
 
 # A dictionary of parameters.