+++ Transpilers/Inline/Inline.Perl.psx.ps1
@@ -31,6 +31,7 @@ $msg = "hello", "hi", "hey", "howdy" | Get-Random
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline)]
+[Management.Automation.CommandInfo]
 $CommandInfo,
 
 # A dictionary of parameters.