+++ Transpilers/Inline/Inline.TOML.psx.ps1
@@ -28,6 +28,7 @@ RandomNumber = """{Get-Random}"""
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline)]
+[Management.Automation.CommandInfo]
 $CommandInfo,
 
 # A dictionary of parameters.